Fix: Daikin ARC433 not working
I kept the original behaviour intact. As far as I can tell, it seems the
original author (@MagicBear), wanted to insert an exception respect to
the standard Daikin protocol, the only problem was that it broke the rest of
the IR versions (included ARC433). The original exception is stille there,
but I cannot test it, maybe the author want to clean up the code in the future.